pyproject.toml
[project]
name = "mcppaylocity"
version = "0.1.0"
description = "A MCP server to fetch data from paylocity endpoint"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.3.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0"
]
[[project.authors]]
name = "MJ Zou"
email = "kelvinzou@outlook.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
mcppaylocity = "mcppaylocity:main"
